requirements.txt
openai
streamlit
pyinstaller
google-cloud-firestore
google-auth-oauthlib